+++ WeightCore/DeviceManagerEntity.cs
@@ -5,7 +5,7 @@ using System;
 using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
-using UICommon;
+using WeightCore.Gui;
 
 namespace WeightCore
 {