+++ WeightCore/MassaK/MkDeviceEntityBackup.cs
@@ -9,7 +9,7 @@
 //using System.Net.Sockets;
 //using System.Threading;
 //using System.Threading.Tasks;
-//using UICommon;
+//using WeightCore.Gui;
 
 //namespace WeightCore.MassaK
 //{