+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.3.020] - 2021-08-09
+## [0.3.020] - 2021-08-10
 ### Changed
 - ScalesUI v.0.5.350
 - WeightCore v.0.5.350